+++ inst/sql/avdth/avdth_elementarycatches.sql
@@ -1,7 +1,7 @@
 SELECT
 	'fr.ird.avdth.entities.data.Activity#'
 		& format(ce.C_BAT, '0000')
-		& '#' 
+		& '#'
 		& YEAR(ce.D_DBQ)
 		& format(MONTH(ce.D_DBQ), '00')
 		& format(DAY(ce.D_DBQ), '00')
@@ -33,14 +33,14 @@ FROM
 	INNER JOIN PAYS p ON b.C_PAYS = p.C_PAYS)
 WHERE
 	a.D_DBQ BETWEEN ?begin_time_period AND ?end_time_period
-	AND p.C_PAYS IN (?flag_codes)
+	AND p.C_ISO3166_A3 IN (?flag_codes)
 	AND tb.C_TYP_B IN (?vessel_type_codes)
 	AND 'fr.ird.avdth.entities.data.Trip#'
 			& format(ce.C_BAT, '0000')
 			& '#'
 			& YEAR(ce.D_DBQ)
 			& format(MONTH(ce.D_DBQ), '00')
-			& format(DAY(ce.D_DBQ), '00') IN (SELECT DISTINCT 
+			& format(DAY(ce.D_DBQ), '00') IN (SELECT DISTINCT
 												'fr.ird.avdth.entities.data.Trip#'
 													& format(a.C_BAT, '0000')
 													& '#'
@@ -53,6 +53,6 @@ WHERE
 												INNER JOIN PAYS p ON b.C_PAYS = p.C_PAYS)
 											WHERE
 												a.D_DBQ BETWEEN ?begin_time_period AND ?end_time_period
-												AND p.C_PAYS IN (?flag_codes)
+												AND p.C_ISO3166_A3 IN (?flag_codes)
 												AND a.C_OCEA IN (?ocean_codes))
 ;